Fix quick view
